Add activeCPUs count for overriding number of CPUs
This change adds functionality for overriding the number of CPUs
returned by `omrsysinfo_get_number_CPUs_by_type` for the
`OMRPORT_CPU_TARGET` type. This is implemented as a port library
global and can be set through a port library function.

Since this functionality fulfills the needs that were previously
satisfied by the `OMRPORT_CPU_ENTITLED` type, this change also
removes entitled CPU count altogether. Active CPUs is now the only
CPU count specified by the user.

This change is to satisfy the need for this functionality, as
discussed in eclipse/openj9#1166.

Issue: eclipse/openj9#1166
Signed-off-by: Mohammad Ali Nikseresht <anikser@ibm.com>